Made improvements, unit tests in progress.
