+++ README.md
@@ -1,12 +1,51 @@
 # BigriverBookstore-api
 
-## Installation
+The jsonapi for the bigriver bookstore
 
-The project requires Python 3 and Pip 3. It was created in a virtual environment using **pipenv**.
+# Plan
+
+Take stories out of the backlog:
+- read it
+- understand it
+- discuss to fill in detalils
+- label it (feature, improvement, bug)
+- estimate it, if it takes more than  day it needs to be split into more stories
+
+# Work
+
+- create a branch for the new feature, name it accordingly 
+- make commits as often as possible, metaly break down the feature into tasks and make sure you commit after each task
+- when possible write a test first (tdd)
+- always write unit tests
+
+# Code push
+
+When the feature is completed:
+- push the code and open a pull request
+- ask for a review
+- push the extra changes if needed
+- squash and merge the changes
+
+
+# Deploy
+
+Deploy via the bulid server, all tests should be green and code coverage should be good. (>95%)
+
+# General notes
+
+Plan the stories with at least one of your colleagues.
+Don't work on this out of the office, you can work on other related projects but keep the `work` in the office.
+
+# Running the project
+
+## Requirements
+* Python >= 3.6
+* Pip
+* pipenv (used to create a virtual environment)
 
 ```sh
 $ pip3 install --user pipenv
-$ cd bookstore
+$ cd BigriverBook-api
 # The following command will install the project's dependencies from the Pipfile.
 $ pipenv install
 # Next, activate the pipenv shell.
@@ -15,20 +54,10 @@ $ pipenv shell
 $ python3 run.py
 ```
 
-## Features
+# Running the specs
 
--   [JSON:API](https://jsonapi.org/) compliant
--   The API currently has 2 endpoints:
-    -   /books/ - JSON response containing data about all books
-    -   /books/book_id - JSON response containing data about a specific book
+TBD
+ 
+# Running linting
 
-## Todos
-
--   Basically, a lot of things
--   Unit Tests
-
-## License
-
-MIT
-
-**Free Software, Hell Yeah!**
+TBD
