+++ api/views.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, jsonify
-from .models import BookSchema
+from .models import Book
+from .schemas import BookSchema
 from .mock_db import db
 
 
@@ -18,12 +19,12 @@ books_bp = Blueprint('books', __name__)
 @books_bp.route('/')
 def get_all():
     books = db['books']
-    data = BookSchema(many=True).dump(books)
+    data = BookSchema(many=True).dump(books).data
     return jsonify_schema_dump(data)
 
 
 @books_bp.route('/<int:book_id>')
 def get_by_id(book_id):
     book = db['books'][book_id - 1]
-    data = BookSchema().dump(book)
+    data = BookSchema().dump(book).data
     return jsonify_schema_dump(data)
